kernel/freertos: Refactor that add LTO optimization feature. [6/9]
PD#SWPL-68240

Problem:
Remove compile error for lto option.

Solution:
Remove compile error for lto option.

Verify:
N/A

+++ bl30/rtos_sdk/kernel/freertos/aml_extend/aml_portable_ext.c
@@ -59,7 +59,7 @@ void vLowPowerSystem(void)
 }
 
 /*-----------------------------------------------------------*/
-uint8_t xPortIsIsrContext(void)
+unsigned int xPortIsIsrContext(void)
 {
 #if CONFIG_ARM64
 	return ullPortInterruptNesting == 0 ? 0 : 1;
